requirements.txt
annotated-types==0.7.0
anyio==4.9.0
atproto==0.0.59
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
libipld==3.0.1
openai==1.70.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
websockets==13.1
